19-1: input parsing and size check
The upper bound on the size of the accepted set is about 2M, so it looks
like the approach I have in mind is feasible.

I'm planning to write a DP-style solution that recusively computes the
accepted set of each node in the DAG, with memoization.

Then we simply use the accepted set to answer all the queries; easy
peasy!